appengine/python-download-kaggle/requirements.txt
wheel
gunicorn
flask
kaggle
google-cloud-bigquery
pylint-quotes
